+++ src/pages/backendconnections/dialog/AddBackendConnectionDialog.js
@@ -1,5 +1,5 @@
 import dataUtils from "@/utils/dataUtils";
-import validationUtils from "../../../../utils/validationUtils";
+import validationUtils from "@/utils/validationUtils";
 
 export default {
     components: {},