+++ src/pages/dataconsumption/dataconsumption/IDSDataConsumptionPage.js
@@ -252,9 +252,9 @@ export default {
             }
         },
 
-        async subscribeToResource() {
+        async subscribeToResource(subscriptionLocation) {
             try {
-                this.$data.subscribeToResourceResponse = await dataUtils.subscribeToResource(this.$data.recipientId, this.$data.selectedResource["@id"]);
+                this.$data.subscribeToResourceResponse = await dataUtils.subscribeToResource(this.$data.recipientId, this.$data.selectedResource["@id"], subscriptionLocation);
             } catch (error) {
                 errorUtils.showError(error, "subscribe to Resource");
             }
@@ -305,14 +305,28 @@ export default {
             this.$data.selectedArtifacts = representation["ids:instance"];
         },
 
-        requestArtifact(item) {
-            this.$refs.artifactDialog.show(this.$data.selectedResource["ids:contractOffer"][0]["ids:permission"], this.$data.selectedResource["ids:standardLicense"]["@id"], item, this.clickAcceptContract);
+        async requestArtifact(item) {
+            let configuration = await dataUtils.getConnectorConfiguration();
+            let subscriptionLocations = [];
+            subscriptionLocations.push({
+                display: configuration.endpoint,
+                value: configuration.endpoint
+            });
+            for (let route of this.$data.routes) {
+                subscriptionLocations.push({
+                    display: route.description,
+                    value: route.selfLink
+                });
+            }
+            this.$refs.artifactDialog.show(this.$data.selectedResource["ids:contractOffer"][0]["ids:permission"], this.$data.selectedResource["ids:standardLicense"]["@id"], item, subscriptionLocations, this.clickAcceptContract);
         },
 
-        clickAcceptContract(artifact) {
+        clickAcceptContract(artifact, subscribe, subscriptionLocation) {
             this.$data.selectedIdsArtifact = artifact;
             this.requestContract();
-            this.subscribeToResource();
+            if (subscribe) {
+                this.subscribeToResource(subscriptionLocation);
+            }
         },
         async dispatchViaRoutes() {
             this.$root.$emit('showBusyIndicator', true);