+++ src/pages/subscriptions/SubscriptionsPage.vue
@@ -1,0 +1,3 @@
+<template src="./SubscriptionsPage.html"></template>
+<script src="./SubscriptionsPage.js"></script>
+<style src="./SubscriptionsPage.css"></style>